Filled some methods in CodeObjectBuilder

+++ Sources/Compiler/Code object/CodeObjectBuilder+Function.swift
@@ -4,31 +4,38 @@ import Bytecode
 
 extension CodeObjectBuilder {
 
-  /// Append a `makeFunction` instruction to currently filled code object.
-  public func emitMakeFunction(argumentCount: UInt8, location: SourceLocation) throws {
+  /// Append a `makeFunction` instruction to code object.
+  public func emitMakeFunction(argumentCount: Int,
+                               location: SourceLocation) throws {
     // try self.emit(.makeFunction, location: location)
     throw self.unimplemented()
   }
 
-  /// Append a `callFunction` instruction to currently filled code object.
-  public func emitCallFunction(argumentCount: UInt8, location: SourceLocation) throws {
+  /// Append a `callFunction` instruction to code object.
+  public func emitCallFunction(argumentCount: Int,
+                               location: SourceLocation) throws {
+    // ADDOP_I(c, CALL_FUNCTION, n + argsCount);
     // try self.emit(.callFunction, location: location)
     throw self.unimplemented()
   }
 
-  /// Append a `callFunctionKw` instruction to currently filled code object.
-  public func emitCallFunctionKw(argumentCount: UInt8, location: SourceLocation) throws {
+  /// Append a `callFunctionKw` instruction to code object.
+  public func emitCallFunctionKw(argumentCount: Int,
+                                 location: SourceLocation) throws {
+    // ADDOP_I(c, CALL_FUNCTION_KW, n + argsCount + keywordCount);
     // try self.emit(.callFunctionKw, location: location)
     throw self.unimplemented()
   }
 
-  /// Append a `callFunctionEx` instruction to currently filled code object.
-  public func emitCallFunctionEx(hasKeywordArguments: Bool, location: SourceLocation) throws {
+  /// Append a `callFunctionEx` instruction to code object.
+  public func emitCallFunctionEx(hasKeywordArguments: Bool,
+                                 location: SourceLocation) throws {
+    // ADDOP_I(c, CALL_FUNCTION_EX, nsubkwargs > 0);
     // try self.emit(.callFunctionEx, location: location)
     throw self.unimplemented()
   }
 
-  /// Append a `returnValue` instruction to currently filled code object.
+  /// Append a `returnValue` instruction to code object.
   public func emitReturnValue(location: SourceLocation) throws {
     try self.emit(.returnValue, location: location)
   }
